Move client tests out of __init__.py file to ensure they run
